requirements.txt
paho-mqtt
pycryptodome
xtea
